Move to MVVM architecture  . Create ViewModel with app logic  . Create FreeWriteFileManager to manage all the file system logic  . AIChatManager to manage ai redirection  . New Models to clean logic  . Extract UI to certain componentes and view to desacoplate the logic and have clean ContentView
