+++ freewrite/Helpers/NSView+Extension.swift
@@ -1,0 +1,39 @@
+//
+//  NSView+Extension.swift
+//  freewrite
+//
+//  Created by Gaspar Dolcemascolo on 18-04-25.
+//
+
+import Foundation
+import AppKit
+
+// Add helper extension to find NSTextView
+extension NSView {
+    func findTextView() -> NSView? {
+        if self is NSTextView {
+            return self
+        }
+        for subview in subviews {
+            if let textView = subview.findTextView() {
+                return textView
+            }
+        }
+        return nil
+    }
+}
+
+// Add helper extension for finding subviews of a specific type
+extension NSView {
+    func findSubview<T: NSView>(ofType type: T.Type) -> T? {
+        if let typedSelf = self as? T {
+            return typedSelf
+        }
+        for subview in subviews {
+            if let found = subview.findSubview(ofType: type) {
+                return found
+            }
+        }
+        return nil
+    }
+}
