+++ freewrite/Managers/AIChatManager.swift
@@ -1,0 +1,60 @@
+//
+//  AIChatManager.swift
+//  freewrite
+//
+//  Created by Gaspar Dolcemascolo on 18-04-25.
+//
+
+import Foundation
+import AppKit
+
+final class AIChatManager {
+    static let shared: AIChatManager = AIChatManager()
+    
+    // Add shared prompt constant
+    private let aiChatPrompt = """
+    below is my journal entry. wyt? talk through it with me like a friend. don't therpaize me and give me a whole breakdown, don't repeat my thoughts with headings. really take all of this, and tell me back stuff truly as if you're an old homie.
+    
+    Keep it casual, dont say yo, help me make new connections i don't see, comfort, validate, challenge, all of it. dont be afraid to say a lot. format with markdown headings if needed.
+    
+    do not just go through every single thing i say, and say it back to me. you need to proccess everythikng is say, make connections i don't see it, and deliver it all back to me as a story that makes me feel what you think i wanna feel. thats what the best therapists do.
+    
+    ideally, you're style/tone should sound like the user themselves. it's as if the user is hearing their own tone but it should still feel different, because you have different things to say and don't just repeat back they say.
+    
+    else, start by saying, "hey, thanks for showing me this. my thoughts:"
+        
+    my entry:
+    """
+    
+    private let claudePrompt = """
+    Take a look at my journal entry below. I'd like you to analyze it and respond with deep insight that feels personal, not clinical.
+    Imagine you're not just a friend, but a mentor who truly gets both my tech background and my psychological patterns. I want you to uncover the deeper meaning and emotional undercurrents behind my scattered thoughts.
+    Keep it casual, dont say yo, help me make new connections i don't see, comfort, validate, challenge, all of it. dont be afraid to say a lot. format with markdown headings if needed.
+    Use vivid metaphors and powerful imagery to help me see what I'm really building. Organize your thoughts with meaningful headings that create a narrative journey through my ideas.
+    Don't just validate my thoughts - reframe them in a way that shows me what I'm really seeking beneath the surface. Go beyond the product concepts to the emotional core of what I'm trying to solve.
+    Be willing to be profound and philosophical without sounding like you're giving therapy. I want someone who can see the patterns I can't see myself and articulate them in a way that feels like an epiphany.
+    Start with 'hey, thanks for showing me this. my thoughts:' and then use markdown headings to structure your response.
+    
+    Here's my journal entry:
+    """
+    
+    func openChatGPT(text: String) {
+        let trimmedText = text.trimmingCharacters(in: .whitespacesAndNewlines)
+        let fullText = aiChatPrompt + "\n\n" + trimmedText
+        
+        if let encodedText = fullText.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed),
+           let url = URL(string: "https://chat.openai.com/?m=" + encodedText) {
+            NSWorkspace.shared.open(url)
+        }
+    }
+    
+    func openClaude(text: String) {
+        let trimmedText = text.trimmingCharacters(in: .whitespacesAndNewlines)
+        let fullText = claudePrompt + "\n\n" + trimmedText
+        
+        if let encodedText = fullText.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed),
+           let url = URL(string: "https://claude.ai/new?q=" + encodedText) {
+            NSWorkspace.shared.open(url)
+        }
+    }
+}
