+++ freewrite/Managers/FreewriterFileManager.swift
@@ -1,0 +1,337 @@
+//
+//  FreewriterFileManager.swift
+//  freewrite
+//
+//  Created by Gaspar Dolcemascolo on 17-04-25.
+//
+
+import Foundation
+import PDFKit
+import AppKit
+import UniformTypeIdentifiers
+
+protocol FreewriterFileManagerProtocole {
+    func createNewEntry() -> HumanEntry
+    func saveEntry(entry: HumanEntry) -> URL
+    func deleteEntry(entry: HumanEntry) -> UUID?
+    func updatePreviewText(for: HumanEntry) -> URL
+    func loadEntry(entry: HumanEntry) -> URL?
+    func exportEntryAsPDF(entry: HumanEntry, font: String, fontSize: CGFloat) -> Void
+    func loadExistingEntries() -> [EntryWithDate]?
+}
+
+final class FreewriterFileManager: FreewriterFileManagerProtocole {
+    
+    static let shared: FreewriterFileManager = FreewriterFileManager()
+    private let fileManager = FileManager.default
+    
+    private let documentsDirectory: URL = {
+        let directory = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)[0].appendingPathComponent("Freewrite")
+        
+        // Create Freewrite directory if it doesn't exist
+        if !FileManager.default.fileExists(atPath: directory.path) {
+            do {
+                try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
+                print("Successfully created Freewrite directory")
+            } catch {
+                print("Error creating directory: \(error)")
+            }
+        }
+        
+        return directory
+    }()
+    
+    // Modify getDocumentsDirectory to use cached value
+    func getDocumentsDirectory() -> URL {
+        return documentsDirectory
+    }
+    
+    
+    // MARK: - LOAD EXISTING ENTIRES
+    
+    
+    func loadExistingEntries() -> [EntryWithDate]? {
+        let documentsDirectory = getDocumentsDirectory()
+        print("Looking for entries in: \(documentsDirectory.path)")
+        
+        do {
+            let fileURLs = try fileManager.contentsOfDirectory(at: documentsDirectory, includingPropertiesForKeys: nil)
+            let mdFiles = fileURLs.filter { $0.pathExtension == "md" }
+            
+            print("Found \(mdFiles.count) .md files")
+            
+            // Process each file
+            let entriesWithDates = mdFiles.compactMap { fileURL -> EntryWithDate? in
+                let filename = fileURL.lastPathComponent
+                print("Processing: \(filename)")
+                
+                // Extract UUID and date from filename - pattern [uuid]-[yyyy-MM-dd-HH-mm-ss].md
+                guard let uuidMatch = filename.range(of: "\\[(.*?)\\]", options: .regularExpression),
+                      let dateMatch = filename.range(of: "\\[(\\d{4}-\\d{2}-\\d{2}-\\d{2}-\\d{2}-\\d{2})\\]", options: .regularExpression),
+                      let uuid = UUID(uuidString: String(filename[uuidMatch].dropFirst().dropLast())) else {
+                    print("Failed to extract UUID or date from filename: \(filename)")
+                    return nil
+                }
+                
+                // Parse the date string
+                let dateString = String(filename[dateMatch].dropFirst().dropLast())
+                let dateFormatter = DateFormatter()
+                dateFormatter.dateFormat = "yyyy-MM-dd-HH-mm-ss"
+                
+                guard let fileDate = dateFormatter.date(from: dateString) else {
+                    print("Failed to parse date from filename: \(filename)")
+                    return nil
+                }
+                
+                // Read file contents for preview
+                do {
+                    let content = try String(contentsOf: fileURL, encoding: .utf8)
+                    let preview = content
+                        .replacingOccurrences(of: "\n", with: " ")
+                        .trimmingCharacters(in: .whitespacesAndNewlines)
+                    let truncated = preview.isEmpty ? "" : (preview.count > 30 ? String(preview.prefix(30)) + "..." : preview)
+                    
+                    // Format display date
+                    dateFormatter.dateFormat = "MMM d"
+                    let displayDate = dateFormatter.string(from: fileDate)
+                    
+                    let entry = HumanEntry(id: uuid, date: displayDate, filename: filename, previewText: truncated)
+                    
+                    return EntryWithDate(entry: entry, date: fileDate, content: content)
+                    
+                } catch {
+                    print("Error reading file: \(error)")
+                    return nil
+                }
+            }
+            return entriesWithDates
+        } catch {
+            print("Error reading file: \(error)")
+            return nil
+        }
+    }
+    
+    // MARK: - UPDATE PREVIEW
+    
+    func updatePreviewText(for entry: HumanEntry) -> URL {
+        let documentsDirectory = getDocumentsDirectory()
+        let fileURL = documentsDirectory.appendingPathComponent(entry.filename)
+        return fileURL
+    }
+    
+    
+    // MARK: - SAVE ENTRY
+    
+    func saveEntry(entry: HumanEntry) -> URL {
+        let documentsDirectory = getDocumentsDirectory()
+        let fileURL = documentsDirectory.appendingPathComponent(entry.filename)
+        return fileURL
+    }
+    
+    
+    // MARK: - CREATE NEW ENTRY
+    
+    func createNewEntry() -> HumanEntry {
+        let newEntry = HumanEntry.createNew()
+        return newEntry
+    }
+    
+    
+    // MARK: - DELETE ENTRY
+    
+    func deleteEntry(entry: HumanEntry) -> UUID? {
+        // Delete the file from the filesystem
+        let documentsDirectory = getDocumentsDirectory()
+        let fileURL = documentsDirectory.appendingPathComponent(entry.filename)
+        
+        do {
+            try fileManager.removeItem(at: fileURL)
+            print("Successfully deleted file: \(entry.filename)")
+            return entry.id
+        } catch {
+            print("Error deleting file: \(error)")
+        }
+        return nil
+    }
+    
+    
+    // MARK: - LOAD ENTRY
+    
+    func loadEntry(entry: HumanEntry) -> URL? {
+        let documentsDirectory = getDocumentsDirectory()
+        let fileURL = documentsDirectory.appendingPathComponent(entry.filename)
+        
+        if fileManager.fileExists(atPath: fileURL.path) {
+            return fileURL
+        }
+        
+        return nil
+    }
+    
+    
+    // MARK: - EXPORT AS PDF
+    
+    func exportEntryAsPDF(entry: HumanEntry, font: String, fontSize: CGFloat) {
+        let documentsDirectory = getDocumentsDirectory()
+        let fileURL = documentsDirectory.appendingPathComponent(entry.filename)
+        
+        do {
+            // Read the content of the entry
+            let entryContent = try String(contentsOf: fileURL, encoding: .utf8)
+            
+            // Extract a title from the entry content and add .pdf extension
+            let suggestedFilename = extractTitleFromContent(entryContent, date: entry.date) + ".pdf"
+            
+            // Create save panel
+            let savePanel = NSSavePanel()
+            savePanel.allowedContentTypes = [UTType.pdf]
+            savePanel.nameFieldStringValue = suggestedFilename
+            savePanel.isExtensionHidden = false  // Make sure extension is visible
+            
+            // Show save dialog
+            if savePanel.runModal() == .OK, let url = savePanel.url {
+                // Create PDF data
+                if let pdfData = createPDFFromText(text: entryContent, font: font, fontSize: fontSize) {
+                    try pdfData.write(to: url)
+                    print("Successfully exported PDF to: \(url.path)")
+                }
+            }
+        } catch {
+            print("Error in PDF export: \(error)")
+        }
+    }
+    
+    private func extractTitleFromContent(_ content: String, date: String) -> String {
+        // Clean up content by removing leading/trailing whitespace and newlines
+        let trimmedContent = content.trimmingCharacters(in: .whitespacesAndNewlines)
+        
+        // If content is empty, just use the date
+        if trimmedContent.isEmpty {
+            return "Entry \(date)"
+        }
+        
+        // Split content into words, ignoring newlines and removing punctuation
+        let words = trimmedContent
+            .replacingOccurrences(of: "\n", with: " ")
+            .components(separatedBy: .whitespaces)
+            .filter { !$0.isEmpty }
+            .map { word in
+                word.trimmingCharacters(in: CharacterSet(charactersIn: ".,!?;:\"'()[]{}<>"))
+                    .lowercased()
+            }
+            .filter { !$0.isEmpty }
+        
+        // If we have at least 4 words, use them
+        if words.count >= 4 {
+            return "\(words[0])-\(words[1])-\(words[2])-\(words[3])"
+        }
+        
+        // If we have fewer than 4 words, use what we have
+        if !words.isEmpty {
+            return words.joined(separator: "-")
+        }
+        
+        // Fallback to date if no words found
+        return "Entry \(date)"
+    }
+    
+    private func createPDFFromText(text: String, font: String, fontSize: CGFloat) -> Data? {
+        // Letter size page dimensions
+        let pageWidth: CGFloat = 612.0  // 8.5 x 72
+        let pageHeight: CGFloat = 792.0 // 11 x 72
+        let margin: CGFloat = 72.0      // 1-inch margins
+        var lineHeight: CGFloat {
+            let font = NSFont(name: font, size: fontSize) ?? .systemFont(ofSize: fontSize)
+            let defaultLineHeight = getLineHeight(font: font)
+            return (fontSize * 1.5) - defaultLineHeight
+        }
+        
+        // Calculate content area
+        let contentRect = CGRect(
+            x: margin,
+            y: margin,
+            width: pageWidth - (margin * 2),
+            height: pageHeight - (margin * 2)
+        )
+        
+        // Create PDF data container
+        let pdfData = NSMutableData()
+        
+        // Configure text formatting attributes
+        let paragraphStyle = NSMutableParagraphStyle()
+        paragraphStyle.lineSpacing = lineHeight
+        
+        let font = NSFont(name: font, size: fontSize) ?? .systemFont(ofSize: fontSize)
+        let textAttributes: [NSAttributedString.Key: Any] = [
+            .font: font,
+            .foregroundColor: NSColor(red: 0.20, green: 0.20, blue: 0.20, alpha: 1.0),
+            .paragraphStyle: paragraphStyle
+        ]
+        
+        // Trim the initial newlines before creating the PDF
+        let trimmedText = text.trimmingCharacters(in: .whitespacesAndNewlines)
+        
+        // Create the attributed string with formatting
+        let attributedString = NSAttributedString(string: trimmedText, attributes: textAttributes)
+        
+        // Create a Core Text framesetter for text layout
+        let framesetter = CTFramesetterCreateWithAttributedString(attributedString)
+        
+        // Create a PDF context with the data consumer
+        guard let pdfContext = CGContext(consumer: CGDataConsumer(data: pdfData as CFMutableData)!, mediaBox: nil, nil) else {
+            print("Failed to create PDF context")
+            return nil
+        }
+        
+        // Track position within text
+        var currentRange = CFRange(location: 0, length: 0)
+        var pageIndex = 0
+        
+        // Create a path for the text frame
+        let framePath = CGMutablePath()
+        framePath.addRect(contentRect)
+        
+        // Continue creating pages until all text is processed
+        while currentRange.location < attributedString.length {
+            // Begin a new PDF page
+            pdfContext.beginPage(mediaBox: nil)
+            
+            // Fill the page with white background
+            pdfContext.setFillColor(NSColor.white.cgColor)
+            pdfContext.fill(CGRect(x: 0, y: 0, width: pageWidth, height: pageHeight))
+            
+            // Create a frame for this page's text
+            let frame = CTFramesetterCreateFrame(
+                framesetter,
+                currentRange,
+                framePath,
+                nil
+            )
+            
+            // Draw the text frame
+            CTFrameDraw(frame, pdfContext)
+            
+            // Get the range of text that was actually displayed in this frame
+            let visibleRange = CTFrameGetVisibleStringRange(frame)
+            
+            // Move to the next block of text for the next page
+            currentRange.location += visibleRange.length
+            
+            // Finish the page
+            pdfContext.endPage()
+            pageIndex += 1
+            
+            // Safety check - don't allow infinite loops
+            if pageIndex > 1000 {
+                print("Safety limit reached - stopping PDF generation")
+                break
+            }
+        }
+        
+        // Finalize the PDF document
+        pdfContext.closePDF()
+        
+        return pdfData as Data
+    }
+    
+}
