+++ freewrite/Models/Chat.swift
@@ -1,0 +1,13 @@
+//
+//  Chat.swift
+//  freewrite
+//
+//  Created by Gaspar Dolcemascolo on 17-04-25.
+//
+
+import Foundation
+
+enum Chat: String {
+    case chatGTP = "ChatGPT"
+    case claude = "Claude"
+}
