+++ freewrite/Models/EntryWithDate.swift
@@ -1,0 +1,14 @@
+//
+//  EntryWithDate.swift
+//  freewrite
+//
+//  Created by Gaspar Dolcemascolo on 18-04-25.
+//
+
+import Foundation
+
+struct EntryWithDate {
+    let entry: HumanEntry
+    let date: Date
+    let content: String
+}
